Move gender repartition graph to its own module
